Downloader for the Purchase Orders of your Amazon Vendor Account
This Excel file contains an UI and and a macro runs on Microsoft IE. By the help of that tool, you can automatically download the POs that you will list in the 'Data' sheet. You need to sign in to your Amazon Vendor Account in IE to use it functionally. It can be used for US as well as UK, AU and CA Amazon Vendor accounts.